UseEffect moved to hooks folder
